Standardize lookup types
